chore: add CODEOWNERS
Code owners are automatically requested for review when someone opens a pull request
that modifies code that they own.

FYI and outside of this commit:
In addition a team (e.g. cloud-sdk-ios-fiori-admin and cloud-sdk-ios-fiori-team) can
leverage code review assignments (it's a setting in GitHub)
Any time a team has been requested to review a pull request, the team is removed as
reviewer and a specified subset of team members are assigned in the team's place.
Code review assignments allow you to decide whether the whole team or just a subset
of team members are notified when a team is requested for review.