ci: being able to trigger sonar scan for pull requests from forks
use pull_request_target

GitHub added a new pull_request_target event, which behaves
 in an almost identical way to the pull_request event with the
same set of filters and payload. However, instead of running
against the workflow and code from the merge commit, the event
runs against the workflow and code from the base of the pull
request. This means the workflow is running from a trusted source
and is given access to a read/write token as well as secrets
enabling the maintainer to safely comment on or label a pull request.

https://github.blog/2020-08-03-github-actions-improvements-for-fork-and-pull-request-workflows/