chore: make required modules for FioriSwiftUI target/library also available for MacCatalyst (#670)
I observed this issue when trying to use FioriSwiftUI package in a framework and using
FioriSwiftUI umbrella library. The framework, when compiling for MacCatalyst, complains
for `import FioriThemeManager`.

The conditional target handling, introduced for supporting visionOS, requires us to specify
the macCatalyst platform. A regression issue but no customer noticed that because we
didn't create a new release yet.